Fixed Windows build of dependencies

+++ _Libs/_sources/build_mingw.sh
@@ -5,4 +5,12 @@ echo $InstallTo
 source ./___build_script.sh
 cp /mingw/bin/libgcc_s_dw2-1.dll    $InstallTo/bin
 cp /mingw/bin/libstdc++-6.dll 	    $InstallTo/bin
-cp /mingw/bin/libwinpthread-1.dll   $InstallTo/bin
+if [ -f /mingw/bin/libwinpthread-1.dll ]; then
+	cp /mingw/bin/libwinpthread-1.dll   $InstallTo/bin
+fi
+if [ -f /mingw/bin/pthreadGC-3.dll ]; then
+	cp /mingw/bin/pthreadGC-3.dll $InstallTo/bin
+fi
+if [ -f /mingw/bin/libmingwex-0.dll ]; then
+	cp /mingw/bin/libmingwex-0.dll $InstallTo/bin
+fi
